Create 'abstract' implementation of ndarray class
